requirements.txt
opencv-python
numpy
pika
moviepy
ultralytics
python-dotenv
